requirements.txt
absl-py==0.8.1
altgraph==0.16.1
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cv==1.0.0.dev8
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
distro==1.4.0
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
future==0.18.0
gast==0.2.2
get==2019.4.13
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.20
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-core==4.5.0
jupyterlab==1.0.9
jupyterlab-server==1.0.6
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Kivy==1.11.1
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.17
kivy-deps.sdl2==0.1.22
Kivy-examples==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.1
numpy==1.16.4
opencv-python==4.1.0.25
openpyxl==2.6.2
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pdf2image==1.11.0
pdfminer==20191020
pefile==2019.4.18
pickleshare==0.7.5
Pillow==6.0.0
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
public==2019.4.13
PyAutoGUI==0.9.50
pycodestyle==2.5.0
pycodestyle-magic==0.5
pycparser==2.19
pycryptodome==3.9.0
pygame==1.9.6
PyGetWindow==0.0.8
Pygments==2.4.2
PyInstaller==3.5
pylint==2.3.1
PyMsgBox==1.0.8
pynput==1.4.2
pyparsing==2.4.0
PyPDF2==1.26.0
pyperclip==1.8.0
pypiwin32==223
PyRect==0.1.4
pyrsistent==0.15.4
PyScreeze==0.1.26
pytesseract==0.3.4
python-dateutil==2.8.0
PyTweening==1.0.3
pytz==2019.3
PyWavelets==1.0.3
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.0
Send2Trash==1.5.0
serious==1.0.0.dev20
six==1.12.0
sounddevice==0.3.13
SoundFile==0.10.2
soupsieve==1.9.3
tabula==1.0.5
tabula-py==1.4.2
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tika==1.23
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.25.3
virtualenv==16.7.5
Wand==0.5.8
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.1
wxPython==4.0.6
